requirements.txt
aniso8601==9.0.1
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
clickclick==20.10.2
connexion==2.14.2
Flask==2.2.5
Flask-Cors==3.0.10
flask-restplus==0.13.0
greenlet==2.0.2
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
packaging==23.1
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.15
swagger-ui==0.1.2
swagger-ui-bundle==0.0.9
typing_extensions==4.6.2
urllib3==2.0.2
Werkzeug==2.2.3
